Relocate path-utils from src/shared to src/plugin/src/utils for plugin-only usage
